More robust attribute handling, closes #14
This patch addresses cases where:

 - An attribute value is the empty string and quotes are not used
 as in `value=${xyz}`
 - An attribute value is a function and quotes _are_ used
 as in `onclick="${onclick}"` (this works in hyperx)
 - An object attribute value contains special characters as in
 `${{ value: 'with"quotes' }}`

To support the first two cases, it checks whether the character right
before an interpolated value is `=`. If so, it adds quotes.

The benchmarks take a significant hit, I guess because of this
additional check. They are up from 290ish ms to 310ish ms on my
machine for 10k renders. Something like an 8% slowdown. I think the
correctness is worth it though, especially doing `value=${xyz}` is quite
common.